move functionalities out of the `batch` modules to the `bank_lines` and the `utils`
