+++ src/dfastbe/batch.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Copyright (C) 2020 Stichting Deltares.
 
@@ -28,8 +27,6 @@ This file is part of D-FAST Bank Erosion: https://github.com/Deltares/D-FAST_Ban
 """
 
 from typing import Tuple, List
-
-import time
 import dfastbe.kernel
 import dfastbe.support
 import dfastbe.io
@@ -40,242 +37,11 @@ import shapely
 import numpy
 import matplotlib
 import configparser
+from dfastbe.utils import timedlogger, adjust_filenames
+from dfastbe.kernel import get_zoom_extends, get_bbox
 
 FIRST_TIME: float
 LAST_TIME: float
-
-
-def banklines(filename: str = "dfastbe.cfg") -> None:
-    """
-    Run the bank line detection analysis using a configuration specified by file name.
-
-    Arguments
-    ---------
-    filename : str
-        Name of the configuration file.
-    """
-    # read configuration file
-    timedlogger("reading configuration file ...")
-    config = dfastbe.io.read_config(filename)
-    rootdir, config = adjust_filenames(filename, config)
-    banklines_core(config, rootdir, False)
-
-
-def adjust_filenames(
-    filename: str, config: configparser.ConfigParser
-) -> Tuple[str, configparser.ConfigParser]:
-    """
-    Convert all paths to relative to current working directory.
-
-    Arguments
-    ---------
-    filename : str
-        Name of the configuration file.
-    config : configparser.ConfigParser
-        Analysis configuration settings.
-    
-    Returns
-    -------
-    rootdir : str
-        Location of configuration file relative to current working directory.
-    config : configparser.ConfigParser
-        Analysis configuration settings using paths relative to current working directory.
-    """
-    rootdir = os.path.dirname(filename)
-    cwd = os.getcwd()
-    config = config_to_absolute_paths(rootdir, config)
-    config = config_to_relative_paths(cwd, config)
-    rootdir = dfastbe.io.relative_path(cwd, rootdir)
-
-    return rootdir, config
-
-
-def banklines_core(config: configparser.ConfigParser, rootdir: str, gui: bool) -> None:
-    """
-    Run the bank line detection analysis for a specified configuration.
-
-    Arguments
-    ---------
-    config : configparser.ConfigParser
-        Analysis configuration settings.
-    rootdir : str
-        Root folder for the analysis (may be relative to current work directory).
-    gui : bool
-        Flag indicating whether this routine is called from the GUI.
-    """
-    timedlogger("-- start analysis --")
-
-    dfastbe.io.log_text(
-        "header_banklines",
-        dict={
-            "version": dfastbe.__version__,
-            "location": "https://github.com/Deltares/D-FAST_Bank_Erosion",
-        },
-    )
-    dfastbe.io.log_text("-")
-
-    # check output dir for bank lines
-    bankdir = dfastbe.io.config_get_str(config, "General", "BankDir")
-    dfastbe.io.log_text("bankdir_out", dict={"dir": bankdir})
-    if os.path.exists(bankdir):
-        dfastbe.io.log_text("overwrite_dir", dict={"dir": bankdir})
-    else:
-        os.makedirs(bankdir)
-
-    # set plotting flags
-    plotting = dfastbe.io.config_get_bool(config, "General", "Plotting", True)
-    if plotting:
-        saveplot = dfastbe.io.config_get_bool(config, "General", "SavePlots", True)
-        saveplot_zoomed = dfastbe.io.config_get_bool(config, "General", "SaveZoomPlots", True)
-        zoom_km_step = dfastbe.io.config_get_float(config, "General", "ZoomStepKM", 1.0)
-        if zoom_km_step < 0.01:
-            saveplot_zoomed = False
-        closeplot = dfastbe.io.config_get_bool(config, "General", "ClosePlots", False)
-    else:
-        saveplot = False
-        saveplot_zoomed = False
-        closeplot = False
-    
-    # as appropriate check output dir for figures and file format
-    if saveplot:
-        figdir = dfastbe.io.config_get_str(
-            config, "General", "FigureDir", rootdir + os.sep + "figure"
-        )
-        dfastbe.io.log_text("figure_dir", dict={"dir": figdir})
-        if os.path.exists(figdir):
-            dfastbe.io.log_text("overwrite_dir", dict={"dir": figdir})
-        else:
-            os.makedirs(figdir)
-        plot_ext = dfastbe.io.config_get_str(config, "General", "FigureExt", ".png")
-
-    # read chainage path
-    xykm = dfastbe.io.config_get_xykm(config)
-
-    # clip the chainage path to the range of chainages of interest
-    kmbounds = dfastbe.io.config_get_kmbounds(config)
-    dfastbe.io.log_text("clip_chainage", dict={"low": kmbounds[0], "high": kmbounds[1]})
-    xykm = dfastbe.io.clip_path_to_kmbounds(xykm, kmbounds)
-    xykm_numpy = numpy.array(xykm)
-    xy_numpy = xykm_numpy[:, :2]
-
-    # read bank search lines
-    max_river_width = 1000
-    search_lines = dfastbe.io.config_get_search_lines(config)
-    search_lines, maxmaxd = dfastbe.support.clip_search_lines(
-        search_lines, xykm, max_river_width
-    )
-    n_searchlines = len(search_lines)
-
-    # convert search lines to bank polygons
-    dlines = dfastbe.io.config_get_bank_search_distances(config, n_searchlines)
-    bankareas = dfastbe.support.convert_search_lines_to_bank_polygons(
-        search_lines, dlines
-    )
-
-    # determine whether search lines are located on left or right
-    to_right = [True] * n_searchlines
-    for ib in range(n_searchlines):
-        to_right[ib] = dfastbe.support.on_right_side(
-            numpy.array(search_lines[ib]), xy_numpy
-        )
-
-    # get simulation file name
-    simfile = dfastbe.io.config_get_simfile(config, "Detect", "")
-
-    # get critical water depth used for defining bank line (default = 0.0 m)
-    h0 = dfastbe.io.config_get_float(config, "Detect", "WaterDepth", default=0)
-
-    # read simulation data and drying flooding threshold dh0
-    dfastbe.io.log_text("-")
-    dfastbe.io.log_text("read_simdata", dict={"file": simfile})
-    dfastbe.io.log_text("-")
-    sim, dh0 = dfastbe.io.read_simdata(simfile)
-    dfastbe.io.log_text("-")
-
-    # increase critical water depth h0 by flooding threshold dh0
-    h0 = h0 + dh0
-
-    # clip simulation data to boundaries ...
-    dfastbe.io.log_text("clip_data")
-    sim = dfastbe.support.clip_simdata(sim, xykm, maxmaxd)
-
-    # derive bank lines (getbanklines)
-    dfastbe.io.log_text("identify_banklines")
-    banklines = dfastbe.support.get_banklines(sim, h0)
-    banklines.to_file(bankdir + os.sep + "raw_detected_bankline_fragments.shp")
-    geopandas.GeoSeries(bankareas).to_file(bankdir + os.sep + "bank_areas.shp")
-
-    # clip the set of detected bank lines to the bank areas
-    dfastbe.io.log_text("simplify_banklines")
-    bank = [None] * n_searchlines
-    clipped_banklines = [None] * n_searchlines
-    for ib, bankarea in enumerate(bankareas):
-        dfastbe.io.log_text("bank_lines", dict={"ib": ib + 1})
-        clipped_banklines[ib] = dfastbe.support.clip_bank_lines(banklines, bankarea)
-        bank[ib] = dfastbe.support.sort_connect_bank_lines(
-            clipped_banklines[ib], xykm, to_right[ib]
-        )
-    geopandas.GeoSeries(clipped_banklines).to_file(
-        bankdir + os.sep + "bankline_fragments_per_bank_area.shp"
-    )
-    dfastbe.io.log_text("-")
-
-    # save bankfile
-    bankname = dfastbe.io.config_get_str(config, "General", "BankFile", "bankfile")
-    bankfile = bankdir + os.sep + bankname + ".shp"
-    dfastbe.io.log_text("save_banklines", dict={"file": bankfile})
-    geopandas.GeoSeries(bank).to_file(bankfile)
-
-    if plotting:
-        dfastbe.io.log_text("=")
-        dfastbe.io.log_text("create_figures")
-        ifig = 0
-        bbox = get_bbox(xykm_numpy)
-
-        if saveplot_zoomed:
-            bank_crds: List[numpy.ndarray] = []
-            bank_km: List[numpy.ndarray] = []
-            for ib in range(n_searchlines):
-                bcrds_numpy = numpy.array(bank[ib])
-                km_numpy = dfastbe.support.project_km_on_line(bcrds_numpy, xykm_numpy)
-                bank_crds.append(bcrds_numpy)
-                bank_km.append(km_numpy)
-            kmzoom, xyzoom = get_zoom_extends(kmbounds[0], kmbounds[1], zoom_km_step, bank_crds, bank_km)
-
-        fig, ax = dfastbe.plotting.plot_detect1(
-            bbox,
-            xykm_numpy,
-            bankareas,
-            bank,
-            sim["facenode"],
-            sim["nnodes"],
-            sim["x_node"],
-            sim["y_node"],
-            sim["h_face"],
-            1.1 * sim["h_face"].max(),
-            "x-coordinate [m]",
-            "y-coordinate [m]",
-            "water depth and detected bank lines",
-            "water depth [m]",
-            "bank search area",
-            "detected bank line"
-        )
-        if saveplot:
-            ifig = ifig + 1
-            figbase = figdir + os.sep + str(ifig) + "_banklinedetection"
-            if saveplot_zoomed:
-                dfastbe.plotting.zoom_xy_and_save(fig, ax, figbase, plot_ext, xyzoom, scale=1)
-            figfile = figbase + plot_ext
-            dfastbe.plotting.savefig(fig, figfile)
-
-    if plotting:
-        if closeplot:
-            matplotlib.pyplot.close("all")
-        else:
-            matplotlib.pyplot.show(block=not gui)
-
-    dfastbe.io.log_text("end_banklines")
-    timedlogger("-- stop analysis --")
 
 
 def bankerosion(filename="dfastbe.cfg") -> None:
@@ -403,7 +169,7 @@ def bankerosion_core(
     dfastbe.io.log_text("derive_topology")
     fn = sim["facenode"]
     nnodes = sim["nnodes"]
-    en, ef, fe, boundary_edge_nrs = derive_topology_arrays(fn, nnodes)
+    en, ef, fe, boundary_edge_nrs = _derive_topology_arrays(fn, nnodes)
 
     # read chainage path
     xykm = dfastbe.io.config_get_xykm(config)
@@ -423,8 +189,8 @@ def bankerosion_core(
     # map bank lines to mesh cells
     dfastbe.io.log_text("intersect_bank_mesh")
     bankline_faces = [None] * n_banklines
-    xf = masked_index(sim["x_node"], fn)
-    yf = masked_index(sim["y_node"], fn)
+    xf = _masked_index(sim["x_node"], fn)
+    yf = _masked_index(sim["y_node"], fn)
     xe = sim["x_node"][en]
     ye = sim["y_node"][en]
     bank_crds = []
@@ -847,7 +613,7 @@ def bankerosion_core(
                 # bankheight = maximum bed elevation per cell
                 if sim["zb_location"] == "node":
                     zb = sim["zb_val"]
-                    zb_all_nodes = masked_index(zb, fnc[bank_index, :])
+                    zb_all_nodes = _masked_index(zb, fnc[bank_index, :])
                     zb_bank = zb_all_nodes.max(axis=1)
                     if zb_dx > 0.0:
                         if ib == 0:
@@ -1345,7 +1111,7 @@ def bankerosion_core(
     timedlogger("-- end analysis --")
 
 
-def masked_index(x0: numpy.array, idx: numpy.ma.masked_array) -> numpy.ma.masked_array:
+def _masked_index(x0: numpy.array, idx: numpy.ma.masked_array) -> numpy.ma.masked_array:
     """
     Index one array by another transferring the mask.
     
@@ -1367,36 +1133,7 @@ def masked_index(x0: numpy.array, idx: numpy.ma.masked_array) -> numpy.ma.masked
     return x1
 
 
-def get_bbox(
-    xykm: numpy.ndarray, buffer: float = 0.1
-) -> Tuple[float, float, float, float]:
-    """
-    Derive the bounding box from a line.
-    
-    Arguments
-    ---------
-    xybm : numpy.ndarray
-        An N x M array containing x- and y-coordinates as first two M entries
-    buffer : float
-        Buffer fraction surrounding the tight bounding box
-    
-    Results
-    -------
-    bbox : Tuple[float, float, float, float]
-        Tuple bounding box consisting of [min x, min y, max x, max y)
-    """
-    x = xykm[:, 0]
-    y = xykm[:, 1]
-    xmin = x.min()
-    ymin = y.min()
-    xmax = x.max()
-    ymax = y.max()
-    d = buffer * max(xmax - xmin, ymax - ymin)
-    bbox = (xmin - d, ymin - d, xmax + d, ymax + d)
-    return bbox
-
-
-def derive_topology_arrays(
+def _derive_topology_arrays(
     fn: numpy.ndarray, n_nodes: numpy.ndarray
 ) -> Tuple[numpy.ndarray, numpy.ndarray, numpy.ndarray, numpy.ndarray]:
     """
@@ -1485,317 +1222,3 @@ def derive_topology_arrays(
     ef[edge_nr[equal_to_previous], 1] = face_nr[equal_to_previous]
 
     return en, ef, fe, boundary_edge_nrs
-
-
-def timedlogger(label: str) -> None:
-    """
-    Write message with time information.
-
-    Arguments
-    ---------
-    label : str
-        Message string.
-    """
-    time, diff = timer()
-    print(time + diff + label)
-
-
-def timer() -> Tuple[str, str]:
-    """
-    Return text string representation of time since previous call.
-
-    The routine uses the global variable LAST_TIME to store the time of the
-    previous call.
-
-    Arguments
-    ---------
-    None
-
-    Returns
-    -------
-    time_str : str
-        String representing duration since first call.
-    diff_str : str
-        String representing duration since previous call.
-    """
-    global FIRST_TIME
-    global LAST_TIME
-    new_time = time.time()
-    if "LAST_TIME" in globals():
-        time_str = "{:6.2f} ".format(new_time - FIRST_TIME)
-        diff_str = "{:6.2f} ".format(new_time - LAST_TIME)
-    else:
-        time_str = "   0.00"
-        diff_str = "       "
-        FIRST_TIME = new_time
-    LAST_TIME = new_time
-    return time_str, diff_str
-
-
-def config_to_absolute_paths(
-    rootdir: str, config: configparser.ConfigParser
-) -> configparser.ConfigParser:
-    """
-    Convert a configuration object to contain absolute paths (for editing).
-
-    Arguments
-    ---------
-    rootdir : str
-        The path to be used as base for the absolute paths.
-    config : configparser.ConfigParser
-        Configuration for the D-FAST Bank Erosion analysis with absolute or relative paths.
-
-    Returns
-    -------
-    config1 : configparser.ConfigParser
-        Configuration for the D-FAST Bank Erosion analysis with only absolute paths.
-    """
-    if "General" in config:
-        config = parameter_absolute_path(config, "General", "RiverKM", rootdir)
-        config = parameter_absolute_path(config, "General", "BankDir", rootdir)
-        config = parameter_absolute_path(config, "General", "FigureDir", rootdir)
-
-    if "Detect" in config:
-        config = parameter_absolute_path(config, "Detect", "SimFile", rootdir)
-        i = 0
-        while True:
-            i = i + 1
-            Line = "Line" + str(i)
-            if Line in config["Detect"]:
-                config = parameter_absolute_path(config, "Detect", Line, rootdir)
-            else:
-                break
-
-    if "Erosion" in config:
-        config = parameter_absolute_path(config, "Erosion", "RiverAxis", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Fairway", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "OutputDir", rootdir)
-
-        config = parameter_absolute_path(config, "Erosion", "ShipType", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "VShip", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "NShip", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "NWave", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Draught", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Wave0", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Wave1", rootdir)
-
-        config = parameter_absolute_path(config, "Erosion", "BankType", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "ProtectionLevel", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Slope", rootdir)
-        config = parameter_absolute_path(config, "Erosion", "Reed", rootdir)
-
-        NLevel = dfastbe.io.config_get_int(config, "Erosion", "NLevel", default=0)
-        for i in range(NLevel):
-            istr = str(i + 1)
-            config = parameter_absolute_path(
-                config, "Erosion", "SimFile" + istr, rootdir
-            )
-            config = parameter_absolute_path(
-                config, "Erosion", "ShipType" + istr, rootdir
-            )
-            config = parameter_absolute_path(config, "Erosion", "VShip" + istr, rootdir)
-            config = parameter_absolute_path(config, "Erosion", "NShip" + istr, rootdir)
-            config = parameter_absolute_path(config, "Erosion", "NWave" + istr, rootdir)
-            config = parameter_absolute_path(
-                config, "Erosion", "Draught" + istr, rootdir
-            )
-            config = parameter_absolute_path(config, "Erosion", "Slope" + istr, rootdir)
-            config = parameter_absolute_path(config, "Erosion", "Reed" + istr, rootdir)
-
-    return config
-
-
-def config_to_relative_paths(
-    rootdir: str, config: configparser.ConfigParser
-) -> configparser.ConfigParser:
-    """
-    Convert a configuration object to contain relative paths (for saving).
-
-    Arguments
-    ---------
-    rootdir : str
-        The path to be used as base for the relative paths.
-    config : configparser.ConfigParser
-        Configuration for the D-FAST Bank Erosion analysis with only absolute paths.
-
-    Returns
-    -------
-    config1 : configparser.ConfigParser
-        Updated configuration for D-FAST Bank Erosion analysis with as much as possible relative paths.
-    """
-    if "General" in config:
-        config = parameter_relative_path(config, "General", "RiverKM", rootdir)
-        config = parameter_relative_path(config, "General", "BankDir", rootdir)
-        config = parameter_relative_path(config, "General", "FigureDir", rootdir)
-
-    if "Detect" in config:
-        config = parameter_relative_path(config, "Detect", "SimFile", rootdir)
-        i = 0
-        while True:
-            i = i + 1
-            Line = "Line" + str(i)
-            if Line in config["Detect"]:
-                config = parameter_relative_path(config, "Detect", Line, rootdir)
-            else:
-                break
-
-    if "Erosion" in config:
-        config = parameter_relative_path(config, "Erosion", "RiverAxis", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Fairway", rootdir)
-        config = parameter_relative_path(config, "Erosion", "OutputDir", rootdir)
-
-        config = parameter_relative_path(config, "Erosion", "ShipType", rootdir)
-        config = parameter_relative_path(config, "Erosion", "VShip", rootdir)
-        config = parameter_relative_path(config, "Erosion", "NShip", rootdir)
-        config = parameter_relative_path(config, "Erosion", "NWave", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Draught", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Wave0", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Wave1", rootdir)
-
-        config = parameter_relative_path(config, "Erosion", "BankType", rootdir)
-        config = parameter_relative_path(config, "Erosion", "ProtectionLevel", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Slope", rootdir)
-        config = parameter_relative_path(config, "Erosion", "Reed", rootdir)
-
-        NLevel = dfastbe.io.config_get_int(config, "Erosion", "NLevel", default=0)
-        for i in range(NLevel):
-            istr = str(i + 1)
-            config = parameter_relative_path(
-                config, "Erosion", "SimFile" + istr, rootdir
-            )
-            config = parameter_relative_path(
-                config, "Erosion", "ShipType" + istr, rootdir
-            )
-            config = parameter_relative_path(config, "Erosion", "VShip" + istr, rootdir)
-            config = parameter_relative_path(config, "Erosion", "NShip" + istr, rootdir)
-            config = parameter_relative_path(config, "Erosion", "NWave" + istr, rootdir)
-            config = parameter_relative_path(
-                config, "Erosion", "Draught" + istr, rootdir
-            )
-            config = parameter_relative_path(config, "Erosion", "Slope" + istr, rootdir)
-            config = parameter_relative_path(config, "Erosion", "Reed" + istr, rootdir)
-
-    return config
-
-
-def parameter_absolute_path(
-    config: configparser.ConfigParser, group: str, key: str, rootdir: str
-) -> configparser.ConfigParser:
-    """
-    Convert a parameter value to contain an absolute path.
-
-    Determine whether the string represents a number.
-    If not, try to convert to an absolute path.
-
-    Arguments
-    ---------
-    config : configparser.ConfigParser
-        Configuration for the D-FAST Bank Erosion analysis.
-    group : str
-        Name of the group in the configuration.
-    key : str
-        Name of the key in the configuration.
-    rootdir : str
-        The path to be used as base for the absolute paths.
-
-    Returns
-    -------
-    config1 : configparser.ConfigParser
-        Updated configuration for the D-FAST Bank Erosion analysis.
-    """
-    if key in config[group]:
-        valstr = config[group][key]
-        try:
-            val = float(valstr)
-        except:
-            config[group][key] = dfastbe.io.absolute_path(rootdir, valstr)
-    return config
-
-
-def parameter_relative_path(
-    config: configparser.ConfigParser, group: str, key: str, rootdir: str
-) -> configparser.ConfigParser:
-    """
-    Convert a parameter value to contain a relative path.
-
-    Determine whether the string represents a number.
-    If not, try to convert to a relative path.
-
-    Arguments
-    ---------
-    config : configparser.ConfigParser
-        Configuration for the D-FAST Bank Erosion analysis.
-    group : str
-        Name of the group in the configuration.
-    key : str
-        Name of the key in the configuration.
-    rootdir : str
-        The path to be used as base for the relative paths.
-
-    Returns
-    -------
-    config1 : configparser.ConfigParser
-        Updated configuration for the D-FAST Bank Erosion analysis.
-    """
-    if key in config[group]:
-        valstr = config[group][key]
-        try:
-            val = float(valstr)
-        except:
-            config[group][key] = dfastbe.io.relative_path(rootdir, valstr)
-    return config
-
-
-def get_zoom_extends(km_min: float, km_max: float, zoom_km_step: float, bank_crds: List[numpy.ndarray], bank_km: List[numpy.ndarray]) -> [List[Tuple[float, float]], List[Tuple[float, float, float, float]]]:
-    """
-    Zoom .
-
-    Arguments
-    ---------
-    km_min : float
-        Minimum value for the chainage range of interest.
-    km_max : float
-        Maximum value for the chainage range of interest.
-    zoom_km_step : float
-        Preferred chainage length of zoom box.
-    bank_crds : List[numpy.ndarray]
-        List of N x 2 numpy arrays of coordinates per bank.
-    bank_km : List[numpy.ndarray]
-        List of N numpy arrays of chainage values per bank.
-
-    Returns
-    -------
-    kmzoom : List[Tuple[float, float]]
-        Zoom ranges for plots with chainage along x-axis.
-    xyzoom : List[Tuple[float, float, float, float]]
-        Zoom ranges for xy-plots.
-    """
-
-    zoom_km_bin = (km_min, km_max, zoom_km_step)
-    zoom_km_bnd = dfastbe.kernel.get_km_bins(zoom_km_bin, type=0, adjust=True)
-    eps = 0.1 * zoom_km_step
-
-    kmzoom: List[Tuple[float, float]] = []
-    xyzoom: List[Tuple[float, float, float, float]] = []
-    inf = float('inf')
-    for i in range(len(zoom_km_bnd)-1):
-        km_min = zoom_km_bnd[i] - eps
-        km_max = zoom_km_bnd[i + 1] + eps
-        kmzoom.append((km_min, km_max))
-        
-        xmin = inf
-        xmax = -inf
-        ymin = inf
-        ymax = -inf
-        for ib in range(len(bank_km)):
-            irange = (bank_km[ib] >= km_min) & (bank_km[ib] <= km_max)
-            range_crds = bank_crds[ib][irange, :]
-            x = range_crds[:, 0]
-            y = range_crds[:, 1]
-            xmin = min(xmin, min(x))
-            xmax = max(xmax, max(x))
-            ymin = min(ymin, min(y))
-            ymax = max(ymax, max(y))
-        xyzoom.append((xmin, xmax, ymin, ymax))
-
-    return kmzoom, xyzoom
